+++ package.json
@@ -182,6 +182,7 @@
     "cheerio": "^0.22.0",
     "chokidar-cli": "^1.2.0",
     "concurrently": "^3.4.0",
+    "content-security-policy-parser": "0.1.0",
     "cookie": "^0.3.1",
     "css-loader": "^0.28.1",
     "deepcopy": "^0.6.3",
@@ -200,6 +201,8 @@
     "intl-locales-supported": "1.0.0",
     "jest": "20.0.3",
     "json-loader": "^0.5.4",
+    "mock-express-request": "0.2.0",
+    "mock-express-response": "0.2.0",
     "node-sass": "^4.5.2",
     "piping": "^1.0.0-rc.4",
     "po2json": "^0.4.5",
