+++ src/core/middleware/index.js
@@ -1,0 +1,5 @@
+export { prefixMiddleware } from './prefixMiddleware';
+export { trailingSlashesMiddleware } from './trailingSlash';
+export { csp, frameguard, hsts, getNoScriptStyles } from './security';
+export { serveAssetsLocally } from './staticAssets';
+export { logRequests } from './logRequests';
