+++ src/core/middleware/logRequests.js
@@ -1,0 +1,9 @@
+import log from 'core/logger';
+
+export function logRequests(req, res, next) {
+  const start = new Date();
+  next();
+  const finish = new Date();
+  const elapsed = finish - start;
+  log.info({ req, res, start, finish, elapsed });
+}
